Join + accept/refuse applicant, TODO: MemberList

+++ The Storyteller/Models/MVillage/Building.cs
@@ -5,7 +5,7 @@ namespace The_Storyteller.Models.MVillage
     class Building
     {
         public string Name { get; set; }
-        public int ProprietaryId { get; set; }
+        public ulong ProprietaryId { get; set; }
         public int Level { get; set; }
     }
 }
